+++ history.md
@@ -5,7 +5,15 @@
 # Project History
 
 
-2023-11-26: v0.0.1
+v0.0.20 (2024-05-20)
+---------------------------------------------------------------------
+
+- Added function: FindMany2( Criteria, Projection, Sort, MaxCount, Options )
+- Updated npm library `@liquicode/jsonstor` to `v0.0.20`.
+- Updated npm library `@liquicode/jsongin` to `v0.0.20`.
+
+
+v0.0.1 (2023-11-26)
 ---------------------------------------------------------------------
 
 - Initial release.